fix launch files, remove logging
