+++ launch/phone_cam_delay.py
@@ -1,0 +1,95 @@
+from launch import LaunchDescription
+from launch_ros.actions import ComposableNodeContainer, LoadComposableNodes
+from launch.actions import LogInfo, TimerAction
+from launch_ros.descriptions import ComposableNode
+
+camera_location = "/dev/video0"  
+camera_format = "BGR"
+camera_width = 640
+camera_height = 480
+camera_fps = 30
+
+# bridge
+gst_width = 640
+gst_height = 480
+gst_fps = 30
+mtu = 500
+bitrate = 1200
+bridge_topic = "camera/image_raw"
+
+# delay (s) b/w GStreamerCam and GStreamerBridge
+delay = 3.0
+
+def generate_launch_description():
+    container = ComposableNodeContainer(
+        name='gstreamer_container',
+        namespace='',
+        package='rclcpp_components',
+        executable='component_container_mt',
+        composable_node_descriptions=[],
+        output='screen',
+    )
+
+    gstreamer_cam_node = ComposableNode(
+        package='gstreamer_ros_bridge',
+        plugin='gstreamer_ros_bridge::GStreamerCam',
+        name='gstreamer_cam_node',
+        parameters=[{
+            'camera_location': camera_location,
+            'camera_format': camera_format,
+            'camera_topic': 'camera/image_raw',
+            'camera_width': camera_width,
+            'camera_height': camera_height,
+            'camera_fps': camera_fps,
+            'custom_pipeline': f'v4l2src device={camera_location} ! videoconvert ! videoscale ! video/x-raw,format={camera_format},width={camera_width},height={camera_height},framerate={camera_fps}/1 ! appsink name=sink sync=false drop=true max-buffers=1',
+            'calibration/model': 'plumb_bob',
+            'calibration/D': [0.06378237699334449, -0.06288559054414458, 0.010299844371743075, 0.001592196375327879, 0.0],
+            'calibration/K': [431.34438684381337, 0.0, 318.6896043226821, 0.0, 430.10747116524595, 255.25822327269975, 0.0, 0.0, 1.0],
+            'calibration/R': [1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0],
+            'calibration/P': [440.250732421875, 0.0, 319.36358907070826, 0.0, 0.0, 436.3755187988281, 259.78431637227914, 0.0, 0.0, 0.0, 1.0, 0.0],
+        }]
+    )
+
+    gstreamer_bridge_node = ComposableNode(
+        package='gstreamer_ros_bridge',
+        plugin='gstreamer_ros_bridge::GStreamerBridge',
+        name='gstreamer_bridge_node',
+        parameters=[{
+            'stream_on': True,
+            'gs_ip': "100.64.9.7",
+            'gs_port': "5602",
+            'gst_width': gst_width,
+            'gst_height': gst_height,
+            'gst_fps': gst_fps,
+            'bitrate': 1200,
+            'mtu': 500,
+            'cam_topic': bridge_topic,
+        }]
+    )
+
+    return LaunchDescription([
+        LogInfo(
+            msg='Starting GStreamerCam'
+        ),
+        container,
+        LoadComposableNodes(
+            target_container=container,
+            composable_node_descriptions=[gstreamer_cam_node]
+        ),
+        LogInfo(
+            msg=f'Awaiting GStreamerCam startup for {delay}s'
+        ),
+        TimerAction(
+            period=delay,
+            actions=[
+                LogInfo(msg='Starting GStreamerBridge'),
+                LoadComposableNodes(
+                    target_container=container,
+                    composable_node_descriptions=[gstreamer_bridge_node]
+                ) 
+            ]
+        ),
+        LogInfo(
+            msg='GStreamerBridge started'
+        ),
+    ])
